terraform: extract types to own file
